Mudanças gerais
Mudanças gerais no Sistema

+++ src/Projeto/Bluray.java
@@ -2,7 +2,7 @@ package Projeto;
 
 public class Bluray extends Item{
 	
-	public Bluray(String nome, int valor){
+	public Bluray(String nome, double valor){
 		
 		super(nome, valor);
 	}
