+++ src/Projeto/Item.java
@@ -3,10 +3,10 @@ package Projeto;
 public class Item {
 	
 	private String nome;
-	private int valor; 
+	private double valor; 
 	private boolean emprestado;
 	
-	public Item(String nome, int valor){
+	public Item(String nome, double valor){
 		
 		this.nome = nome;
 		this.valor = valor;
@@ -23,12 +23,12 @@ public class Item {
 		this.nome = nome;
 	}
 
-	public int getValor() {
+	public double getValor() {
 		
 		return this.valor;
 	}
 
-	public void setValor(int valor) {
+	public void setValor(double valor) {
 		
 		this.valor = valor;
 	}
