+++ src/Projeto/JogosEletronicos.java
@@ -3,7 +3,7 @@ package Projeto;
 public class JogosEletronicos extends Item {
 	
 	
-	public JogosEletronicos(String nome, int valor){
+	public JogosEletronicos(String nome, double valor){
 		
 		super(nome, valor);
 	}
